flask-image-analysis/requirements.txt
Flask
opencv-python-headless
numpy
piexif
torch
torchvision
pillow
deepface
werkzeug
gunicorn
